+++ frontend/src/components/FilterDropdown.tsx
@@ -9,9 +9,6 @@ const AllFiltersContainer = styled.div`
   flex-direction: column;
   justify-content: flex-start;
   align-items: flex-start;
-  margin-left: 95px;
-  margin-right: 95px;
-  margin-top: 55px;
   gap: 16px;
 `;
