+++ frontend/src/pages/Home.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useContext, useEffect, useState } from "react";
 import { Helmet } from "react-helmet-async";
 import { useLocation } from "react-router-dom";
 import styled from "styled-components";
@@ -10,15 +10,16 @@ import { NavBar } from "@/components/NavBar";
 import { Page } from "@/components/Page";
 import { UnitCardGrid } from "@/components/UnitCardGrid";
 import { UnitList } from "@/components/UnitList";
+import { DataContext } from "@/contexts/DataContext";
 
 const ButtonsWrapper = styled.div`
   display: flex;
-  flex: row;
   justify-content: end;
-  margin: 0;
-  margin-top: 55px;
-  margin-right: 100px;
-  margin-left: 310px;
+`;
+
+const HeaderText = styled.span`
+  font-family: "Neutraface Text";
+  font-size: 32px;
 `;
 
 const ToggleButtonWrapper = styled.div`
@@ -56,6 +57,7 @@ const ListViewButton = styled(CardViewButton)`
 const SearchStateWrapper = styled.div`
   display: flex;
   flex-direction: row;
+  justify-content: space-between;
 `;
 
 export type FilterContextType = {
@@ -65,6 +67,43 @@ export type FilterContextType = {
 
 export const FiltersContext = React.createContext({} as FilterContextType);
 
+const PropertiesRow = styled.span`
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  color: black;
+  font-family: "Montserrat";
+  font-size: 27px;
+  font-weight: 700;
+  margin-bottom: 30px;
+`;
+
+const PendingButton = styled.div<{ selected: boolean }>`
+  display: flex;
+  width: 164px;
+  height: 55px;
+  align-items: center;
+  justify-content: center;
+  border-radius: ${(props) => (props.selected ? "12px" : "12px 0px 0px 12px")};
+  border: 1px solid ${(props) => (props.selected ? "rgba(162, 61, 4, 0.80)" : "#EEE")};
+  background: ${(props) => (props.selected ? "#B64201" : "#EEE")};
+  color: ${(props) => (props.selected ? "#EEE" : "#2E2E2E")};
+  font-family: Poppins;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 600;
+  line-height: normal;
+  z-index: ${(props) => (props.selected ? 1 : 0)};
+  cursor: pointer;
+`;
+
+const ListingsButton = styled(PendingButton)`
+  border-radius: ${(props) => (props.selected ? "12px" : "0px 12px 12px 0px")};
+  position: relative;
+  left: -10px;
+`;
+
 const HomePageLayout = styled.div`
   display: flex;
   flex-direction: row;
@@ -72,12 +111,14 @@ const HomePageLayout = styled.div`
   height: 100%;
 `;
 
-const FilterPadding = styled.div`
-  min-width: 250px;
-  height: 100%;
+const UnitContent = styled.div`
+  width: 100%;
+  max-height: calc(100vh - 70px);
+  overflow-y: scroll;
+  padding: 70px 60px;
 `;
-
 export function Home() {
+  const dataContext = useContext(DataContext);
   const previousFilters = useLocation().state as FilterParams;
   const [units, setUnits] = useState<Unit[]>([]);
   const [filters, setFilters] = useState<FilterParams>(
@@ -87,6 +128,7 @@ export function Home() {
     },
   );
   const [viewMode, setViewMode] = useState("card");
+  const [pendingSelected, setPendingSelected] = useState<boolean>(filters.approved === "pending");
 
   const fetchUnits = (filterParams: FilterParams) => {
     let query: GetUnitsParams = {
@@ -130,6 +172,11 @@ export function Home() {
       .catch(console.error);
   };
 
+  const refreshUnits = (approved: "pending" | "approved") => {
+    const newFilters = { ...filters, approved };
+    setFilters(newFilters);
+  };
+
   useEffect(() => {
     fetchUnits(filters);
   }, [filters]);
@@ -151,8 +198,7 @@ export function Home() {
         <NavBar page="Home" />
         <HomePageLayout>
           <FilterPanel />
-          <FilterPadding />
-          <div>
+          <UnitContent>
             <SearchStateWrapper>
               <FilterDropdown
                 searchText={filters.search ?? ""}
@@ -181,27 +227,45 @@ export function Home() {
                 </ToggleButtonWrapper>
               </ButtonsWrapper>
             </SearchStateWrapper>
+            <PropertiesRow>
+              {pendingSelected ? (
+                <HeaderText>Pending Approval</HeaderText>
+              ) : (
+                <HeaderText>Available Properties</HeaderText>
+              )}
+              {dataContext.currentUser?.isHousingLocator && (
+                <ButtonsWrapper>
+                  <PendingButton
+                    onClick={() => {
+                      setPendingSelected(true);
+                      refreshUnits("pending");
+                    }}
+                    selected={pendingSelected}
+                  >
+                    Pending Listings
+                  </PendingButton>
+                  <ListingsButton
+                    onClick={() => {
+                      setPendingSelected(false);
+                      refreshUnits("approved");
+                    }}
+                    selected={!pendingSelected}
+                  >
+                    All Listings
+                  </ListingsButton>
+                </ButtonsWrapper>
+              )}
+            </PropertiesRow>
             {viewMode === "card" ? (
               <UnitCardGrid
                 units={units}
                 showPendingUnits={filters.approved === "pending"}
-                refreshUnits={(approved) => {
-                  const newFilters = { ...filters, approved };
-                  fetchUnits(newFilters);
-                  setFilters(newFilters);
-                }}
+                refreshUnits={refreshUnits}
               />
             ) : (
-              <UnitList
-                units={units}
-                showPendingUnits={filters.approved === "pending"}
-                refreshUnits={(approved) => {
-                  const newFilters = { ...filters, approved };
-                  setFilters(newFilters);
-                }}
-              />
+              <UnitList units={units} />
             )}
-          </div>
+          </UnitContent>
         </HomePageLayout>
       </Page>
     </FiltersContext.Provider>
